JPA를 이용한 연동 완료!

+++ src/main/java/com/kh/crud/service/PostService.java
@@ -15,22 +15,37 @@ public class PostService {
     private final PostRepository postRepository;
 
     public Post createPost(Post post) {
+
         return postRepository.save(post);
     }
 
-    public Optional<Post> findById(Long id) {
+    public Optional<Post> getPostById(Long id) {
+
         return postRepository.findById(id);
     }
 
-    public List<Post> findAll() {
+    public List<Post> getAllPost() {
+
         return postRepository.findAll();
     }
 
     public Post updatePost(Post post) {
+
         return postRepository.save(post);
     }
 
-    public void deletePost(Long id) {
-        postRepository.deleteById(id);
+    public boolean deletePost(Long id) {
+        // 존재할수도 아닐수도 있는 id를 조회한다.(optional)
+        Optional<Post> optionalPost = postRepository.findById(id);
+
+        if (optionalPost.isPresent()) {
+            postRepository.deleteById(id); // 삭제
+            return true; // 삭제 성공
+        } else {
+            return false; // 존재하지 않음
+        }
     }
+
+
+
 }
